requirements.txt
flask
numpy
sentence-transformers
transformers
scikit-learn
llama-cpp-python
einops
torch